Reorganize metrics and data parser (#712)
* Reorganize metrics and data parser

* Add missing test file
